Feedback updatee: Move SearchFilter to products folder
