+++ client/src/components/products/ProductsList.js
@@ -1,7 +1,7 @@
 import React, { Component } from "react";
 import axios from "axios";
 import { Link } from "react-router-dom";
-import SearchFilter from "../SearchFilter";
+import SearchFilter from "./SearchFilter";
 import useStyles from "./ProductListStyles";
 import {
   Typography,
